refactor(system)!: should not let user to instantiate the System object
Now `System` is a abstract class, instead of a interface, because I
don't want to encourage user to use lambda expression or instantiate
the `System` object by themselves, this will be messy when they pass Object using
the constructor. So now `World.addStartupSystem` and `World.addSystem` only take
`System.class` as parameter.